리팩토링 : MVC패턴을 고려하여 view 수정
